re-organizes file structure for easier navigation
renames files/folders for consistent casing
