--- src/components/Marketing/_marketing.scss
@@ -1,56 +1,0 @@
-
-body{
-  background: #f6f9fa;
-}
-.marketing{
-  background: linear-gradient(147.75deg, #FFFFFF 34.63%, rgba(255, 255, 255, 0) 90.04%), url("../../assets/hero.jpg");
-  background-size: 100% 100%;
-  background-position: center;
-  background-repeat: no-repeat;
-  @media only screen and (min-width: 1200px){
-    background-size: cover;
-  }
-}
-
-.wave{
-  background-image: url("../../assets/wave.png");
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: 100vw 100%;
-  height: 35vh;
-  width: 100%;
-  position: relative;
-  top: 10vh;
-}
-
-.underline{
-  width: 60%;
-  margin: 0 auto;
-}
-
-.badges{
-  border-radius: 55px;
-  box-shadow: -15px -15px 15px rgba(255, 255, 255, 0.651), 13px 16px 25px rgba(139, 139, 139, 0.2);
-}
-
-.carousel-control-next-icon {
-  background-image : url("data:image/svg+xml;charset=utf8,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='%23ff0000' viewBox='0 0 8 8'%3E%3Cpath d='M1.5 0l-1.5 1.5 2.5 2.5-2.5 2.5 1.5 1.5 4-4-4-4z'/%3E%3C/svg%3E");
-  color: black;
-}
-
-.carousel-control-prev-icon {
-  background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='%23ff0000' width='8' height='8' viewBox='0 0 8 8'%3e%3cpath d='M5.25 0l-4 4 4 4 1.5-1.5L4.25 4l2.5-2.5L5.25 0z'/%3e%3c/svg%3e");
-}
-
-#demo{
-  width: 80vw;
-  max-width: 800px; 
-  border-radius: 55px;
-  box-shadow: -15px -15px 15px rgba(255, 255, 255, 0.651), 13px 16px 25px rgba(139, 139, 139, 0.2);
-  img{
-    width: 80vw;
-    max-width: 800px; 
-    border-radius: 55px;
-  }
-}
-
