--- src/components/on-boarding/Buttons/BackBttn.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-const BackBttn = ({ prevStep }) => {
-  return (
-    <button
-      onClick={() => prevStep()}
-      className="mt-3 btn-secondary rounded p-2 w-100"
-    >
-      Back
-    </button>
-  );
-};
-
-export default BackBttn;
