--- src/components/on-boarding/Buttons/NextBttn.js
@@ -1,38 +1,0 @@
-import Axios from "axios";
-import { data } from "jquery";
-import React from "react";
-import { connect, useSelector } from "react-redux";
-import { login } from "../../../State/Slices/slices";
-
-var jwt_decode = require("jwt-decode");
-
-const NextBttn = ({ handleSubmit, nextStep }) => {
-  const onSubmit = async (data) => {
-    const { username, email, password } = data;
-    Axios.post(`https://nutrivurv-be.herokuapp.com/api/auth/register`, {
-      name: username,
-      email: email,
-      password: password,
-    })
-      .then((res) => {
-        console.log(res.data);
-      })
-      .catch((err) => console.log("error", err));
-    nextStep();
-  };
-
-  return (
-    <>
-      <button
-        data-cy="submit"
-        type="submit"
-        className="btn-primary rounded p-2 w-100 border border-primary"
-        onClick={handleSubmit(onSubmit)}
-      >
-        Continue
-      </button>
-    </>
-  );
-};
-
-export default connect(null)(NextBttn);
