--- src/components/on-boarding/onBoarding-img.js
@@ -1,15 +1,0 @@
-import React from "react";
-import { ReactComponent as GirlComptr } from "../../assets/GirlComptr.svg";
-
-const OnBoardingImg = () => {
-  return (
-    <div>
-      <GirlComptr
-        className=" d-none d-xl-block col img-responsive mt-5 pt-5"
-        alt="Responsive Dietary Preference image"
-      />
-    </div>
-  );
-};
-
-export default OnBoardingImg;
