+++ src/index.js
@@ -5,8 +5,8 @@ import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import { BrowserRouter as Router } from "react-router-dom";
 import App from "./App";
-import { store } from "./State/configureStore";
 import * as serviceWorker from "./serviceWorker";
+import { store } from "./state/configureStore";
 
 ReactDOM.render(
   <React.StrictMode>
